requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
uWSGI>=2.0.19.1,<2.1
